:tada: Restart project for DiscordAuth 4.0.0
