--- src/main/java/fr/xen0xys/discordauth/discord/embeds/CustomEmbed.java
@@ -1,7 +1,0 @@
-package fr.xen0xys.discordauth.discord.embeds;
-
-import net.dv8tion.jda.api.entities.MessageEmbed;
-
-public interface CustomEmbed {
-    MessageEmbed build();
-}
