--- src/main/java/fr/xen0xys/discordauth/plugin/commands/LogoutCommand.java
@@ -1,22 +1,0 @@
-package fr.xen0xys.discordauth.plugin.commands;
-
-import fr.xen0xys.discordauth.DiscordAuth;
-import fr.xen0xys.discordauth.models.Commands;
-import org.bukkit.ChatColor;
-import org.bukkit.command.Command;
-import org.bukkit.command.CommandExecutor;
-import org.bukkit.command.CommandSender;
-import org.bukkit.entity.Player;
-import org.jetbrains.annotations.NotNull;
-
-public class LogoutCommand implements CommandExecutor {
-    @Override
-    public boolean onCommand(@NotNull CommandSender commandSender, @NotNull Command command, @NotNull String s, @NotNull String[] strings) {
-        if(commandSender instanceof Player player){
-            Commands.logout(player);
-        }else{
-            commandSender.sendMessage(ChatColor.RED + DiscordAuth.getLanguage().cannotBeExecutedFromCLI);
-        }
-        return false;
-    }
-}
