--- src/main/java/fr/xen0xys/discordauth/plugin/events/OnBlockBreak.java
@@ -1,17 +1,0 @@
-package fr.xen0xys.discordauth.plugin.events;
-
-import fr.xen0xys.discordauth.DiscordAuth;
-import fr.xen0xys.discordauth.models.User;
-import org.bukkit.event.EventHandler;
-import org.bukkit.event.Listener;
-import org.bukkit.event.block.BlockBreakEvent;
-
-public class OnBlockBreak implements Listener {
-    @EventHandler
-    public void onBlockBreak(BlockBreakEvent e){
-        User user = DiscordAuth.getUsers().get(e.getPlayer().getName());
-        if(user != null && !user.isLogged()){
-            e.setCancelled(true);
-        }
-    }
-}
