--- src/main/java/fr/xen0xys/discordauth/plugin/events/OnPlayerDropItem.java
@@ -1,17 +1,0 @@
-package fr.xen0xys.discordauth.plugin.events;
-
-import fr.xen0xys.discordauth.DiscordAuth;
-import fr.xen0xys.discordauth.models.User;
-import org.bukkit.event.EventHandler;
-import org.bukkit.event.Listener;
-import org.bukkit.event.player.PlayerDropItemEvent;
-
-public class OnPlayerDropItem implements Listener {
-    @EventHandler
-    public void onPlayerDropItem(PlayerDropItemEvent e){
-        User user = DiscordAuth.getUsers().get(e.getPlayer().getName());
-        if(user != null && !user.isLogged()){
-            e.setCancelled(true);
-        }
-    }
-}
