--- src/main/java/fr/xen0xys/discordauth/plugin/events/OnPlayerInteract.java
@@ -1,18 +1,0 @@
-package fr.xen0xys.discordauth.plugin.events;
-
-import fr.xen0xys.discordauth.DiscordAuth;
-import fr.xen0xys.discordauth.models.User;
-import org.bukkit.event.EventHandler;
-import org.bukkit.event.EventPriority;
-import org.bukkit.event.Listener;
-import org.bukkit.event.player.PlayerInteractEvent;
-
-public class OnPlayerInteract implements Listener {
-    @EventHandler(priority = EventPriority.HIGHEST)
-    public void onPlayerInteract(PlayerInteractEvent e){
-        User user = DiscordAuth.getUsers().get(e.getPlayer().getName());
-        if(user != null && !user.isLogged()){
-            e.setCancelled(true);
-        }
-    }
-}
