--- src/main/java/fr/xen0xys/discordauth/plugin/events/OnPlayerKick.java
@@ -1,17 +1,0 @@
-package fr.xen0xys.discordauth.plugin.events;
-
-import fr.xen0xys.discordauth.DiscordAuth;
-import org.bukkit.entity.Player;
-import org.bukkit.event.EventHandler;
-import org.bukkit.event.Listener;
-import org.bukkit.event.player.PlayerKickEvent;
-
-public class OnPlayerKick implements Listener {
-    @EventHandler
-    public void onPlayerKick(PlayerKickEvent e){
-        Player player = e.getPlayer();
-        if(e.reason().toString().equals("Flying is not enabled on this server") && !DiscordAuth.getUsers().get(player.getName()).isLogged()){
-            e.setCancelled(true);
-        }
-    }
-}
