--- src/main/java/fr/xen0xys/discordauth/utils/ConsoleFilter.java
@@ -1,43 +1,0 @@
-package fr.xen0xys.discordauth.utils;
-
-import org.apache.logging.log4j.Level;
-import org.apache.logging.log4j.Marker;
-import org.apache.logging.log4j.core.LogEvent;
-import org.apache.logging.log4j.core.Logger;
-import org.apache.logging.log4j.core.filter.AbstractFilter;
-import org.apache.logging.log4j.message.Message;
-
-public class ConsoleFilter extends AbstractFilter {
-
-    @Override
-    public Result filter(LogEvent event) {
-        return event == null ? Result.NEUTRAL : isLoggable(event.getMessage().getFormattedMessage());
-    }
-
-    @Override
-    public Result filter(Logger logger, Level level, Marker marker, Message msg, Throwable t) {
-        return isLoggable(msg.getFormattedMessage());
-    }
-
-    @Override
-    public Result filter(Logger logger, Level level, Marker marker, String msg, Object... params) {
-        return isLoggable(msg);
-    }
-
-    @Override
-    public Result filter(Logger logger, Level level, Marker marker, Object msg, Throwable t) {
-        return msg == null ? Result.NEUTRAL : isLoggable(msg.toString());
-    }
-
-    private Result isLoggable(String msg) {
-        if (msg != null) {
-            if (msg.contains("issued server command:")) {
-                if (msg.contains("/login") || msg.contains("/l") || msg.contains("/account"))
-                    return Result.DENY;
-            }else if(msg.contains("was kicked for floating too long!")){
-                return Result.DENY;
-            }
-        }
-        return Result.NEUTRAL;
-    }
-}
