--- src/main/java/fr/xen0xys/discordauth/utils/StatusColor.java
@@ -1,20 +1,0 @@
-package fr.xen0xys.discordauth.utils;
-
-import java.awt.*;
-
-public enum StatusColor {
-
-    OK(Color.GREEN),
-    Warning(Color.ORANGE),
-    Error(Color.RED);
-
-    private Color color;
-
-    StatusColor(Color color){
-        this.color = color;
-    }
-
-    public Color getColor() {
-        return color;
-    }
-}
