+++ src/main/resources/bungee.yml
@@ -1,0 +1,6 @@
+author: Xen0Xys
+name: DiscordAuth
+version: 4.0.0-beta
+api-version: 1.19
+main: fr.xen0xys.discordauth.waterfall.DiscordAuthProxy
+database: true
