--- src/main/resources/resources/en_EN.yml
@@ -1,51 +1,0 @@
-# Login:
-login_success: "You are login successful"
-login_request: "Please login (/login)"
-reload_login_request: "DiscordAuth has been reloaded, please re-login (/login)!"
-discord_login: "You are login successful (Discord)"
-forced_login: "You are login successful (Forcelogin)"
-session_login: "You are login successful (Session)"
-need_login: "You need to be logged in to do that"
-player_force_logged: "Player has been force logged!"
-connect_once: "You need to connect once to server to do that!"
-
-# Account:
-account_created_successful: "Your account has been created successful!"
-already_has_account: "You already has an account!"
-account_not_exist: "You don't has account, please create one!"
-no_reaction: "Your account cannot be created because you have not react to the require message!"
-
-# Password:
-updated_password: "Your password has been updated!"
-invalid_password: "Invalid password!"
-
-# Discord or Minecraft chat:
-player_connect: "%s has just been connected"
-player_disconnect: "%s has just been disconnected"
-server_start: ":white_check_mark: **Server has just started**"
-server_stop: ":octagonal_sign: **Server has just stopped**"
-
-# Errors:
-error_occurred: "An error occurred"
-already_connected: "Specified player is already connected!"
-session_end: "Your session could not be renewed!"
-player_not_connected: "You are not connected on the server!"
-cant_be_executed_from_cli: "This command can't be executed from console!"
-unknown_specified_player: "&cUnknown specified player"
-kick_message: "You don't has access to this server!\nIn case of error, please contact engineer!"
-password_define_error: "The password length may be between 5 and 50 chars\nThey may contain only alphanumerical chars and following chars: @$!%*#?&"
-
-# Information
-accept_message: "You have accepted the message, now you can create your account by sending the command **/account create** (if not available in MP, send it in the owner's discord server)"
-action_success: "Action executed with success!"
-
-# Adds
-account_deleted: "Account deleted successful"
-no_account_for_username: "No account found for this username"
-id_for_username: "Account found for this username : %s"
-invalid_username_password: "Username or password is invalid"
-no_identical_passwords: "Passwords are not identical"
-account_deletion_confirm: "Are you sure you want to delete this account?"
-keep_account: "Keep account"
-delete_account: "Delete anyways"
-no_permission: "You do not have permission to use this command"
