--- src/main/resources/resources/fr_FR.yml
@@ -1,70 +1,0 @@
-# Login:
-login_success: "Vous êtes connecté avec succès"
-login_request: "Veuillez vous connecter (/login)"
-reload_login_request: "DiscordAuth à été reload, veuillez vous reconnecter (/login)!"
-discord_login: "Vous êtes connecté avec succès (Discord)"
-forced_login: "Vous êtes connecté avec succès (Forcelogin)"
-session_login: "Vous êtes connecté avec succès (Session)"
-need_login: "Vous avez besoin d'être connecté pour faire cela!"
-player_force_logged: "Le joueur spécifié à été connecté de force!"
-connect_once: "Vous avez besoin de vous connecter une première fois au serveur pour faire cela!"
-
-# Account:
-account_created_successful: "Votre compte à été crée avec succès!"
-already_has_account: "Vous possédez déjà un compte!"
-account_not_exist: "Vous n'avez pas encore de compte, veuillez en créer un!"
-no_reaction: "Votre compte n'a pas pu être crée car vous n'avez pas réagi au message demandé!"
-unauthorized_connection: "Une connection non-autorisé a été détecté avec l'ip: %s\nEnvoyez /ip allow %s pour autoriser cette ip!\nEnvoyez /ip block %s pour bloquer cette ip!"
-
-# Password:
-updated_password: "Votre mot de passe à été mis à jour!"
-invalid_password: "Mot de passe invalide!"
-
-# Discord or Minecraft chat:
-player_connect: "%s vient de se connecter"
-player_disconnect: "%s vient de se déconnecter"
-register_chat: "&9<%s>&r %s"
-unregistered_chat: "&6[Non-enregistré] &9<%s>&r %s"
-server_start: ":white_check_mark: **Le serveur vient de démarer**"
-server_stop: ":octagonal_sign: **Le serveur vient de s'arréter**"
-
-# Errors:
-error_occurred: "Une erreur est survenue"
-already_connected: "Le joueur spécifié est déjà connecté!"
-session_end: "Votre session n'a pas pu être renouvelé"
-player_not_connected: "Vous n'êtes pas connecté sur le serveur!"
-cant_be_executed_from_cli: "Cette commande ne peut pas être exécuté depuis la console!"
-unknown_specified_player: "Le joueur spécifié est introuvable"
-kick_message: "Vous n'avez pas accès à ce serveur!\nEn cas d'erreur, contactez un technicien!"
-password_define_error: "Le mot de passe doit faire entre 6 et 50 caractères\nIl doit aussi contenir une majuscule et une minuscule, ainsi qu'un caractère spécial!"
-
-# Advancements:
-success_complete: "%s a complété le(s) succès suivant(s): %s"
-too_many_advancements: "Trop de succès à montrer"
-
-# IP
-invalid_ip_format: "Format d'IP invalide!"
-no_ip_allowed: "Aucune IP autorisé"
-no_ip_blocked: "Aucune IP bloqué"
-all_ip_allowed: "Toutes les IP sont autorisés"
-all_ip_blocked: "Toutes les IP sont bloqués"
-ip_allowed: "IP autorisés:"
-ip_blocked: "IP bloqués:"
-unauthorized_ip: "Votre ip n'a pas été autorisé par le propriétaire du compte!"
-
-# Information
-accept_message: "Vous avez accepté le message, vous pouvez maintenant créer votre compte en envoyant la commande **/account create** (si elle n'est pas disponible en MP, envoyez-la sur le serveur discord du propriétaire)."
-action_success: "L'action a bien été effectué!"
-allowed: "Autorisés:"
-blocked: "Bloqués:"
-
-# Adds
-account_deleted: "Compte supprimé avec succès"
-no_account_for_username: "Aucun compte n'a été trouvé pour ce nom d'utilisateur"
-id_for_username: "Compte trouvé pour ce nom d'utilisateur : %s"
-invalid_username_password: "Le nom d'utilisateur ou le mot de passe n'est pas valide"
-no_identical_passwords: "Les mots de passe ne sont pas identiques"
-account_deletion_confirm: "Etes-vous sûr de vouloir supprimer ce compte?"
-keep_account: "Garder le compte"
-delete_account: "Supprimer quand même"
-no_permission: "Vous n'avez pas la permission d'utiliser cette commande"
